+++ djpcms/tests/__init__.py
@@ -3,3 +3,4 @@ from djpcms.tests.archive_views import *
 from djpcms.tests.test_navigation import *
 from djpcms.tests.doc_views import *
 from djpcms.tests.uniforms import *
+from autocomplete import *